fix: resolve all 82 mypy type errors across 14 files
- scrolling.py (25→0): Add type annotations for _weakest_sections, _backing_muted,
  _audio_capture, _timing_summary. Fix _toggle_string→toggle_string calls. Add
  _toggle_wait_mode method. Remove invalid adaptive_scoring kwarg. Add
  get_phrase_statistics to NoteMatcher.
- evaluation/runner.py (28→0): Fix **dict unpacking to EvaluationRecord by
  constructing records explicitly. Fix ChordVerification/NoteVerification type
  confusion with isinstance checks.
- evaluation/manifest.py (10→0): Cast int()/float() on object types.
- config.py (6→0): Cast object→str/int in apply_preset. Add isinstance check
  for __dataclass_fields__.
- matcher.py (3→0): Fix ExpectedEventVerifier import path. Add type annotation
  for results list. Add missing return.
- log_frequency.py (2→0): Change midi_to_freq parameter to float.
- analyzer.py (1→0): Fix list[int]→list[float] parameter type.
- verifier.py (1→0): Fix ChordVerification/NoteVerification assignment.
- trajectory.py (1→0): Change midi_to_freq parameter to float.
- latency_calibrator.py (1→0): Add isinstance check for dict.get.
- midi_playback.py (1→0): Add type annotation for Output variable.
- calibration_menu.py (1→0): Add type annotation for LatencyResult.
- loader.py (1→0): Add type annotation for measures list.
- records.py (1→0): Fix tuple(object) type.

mypy: 82 errors → 0. Tests: 1746 passed, 0 regressions. Ruff: all checks passed.

+++ pickhero/audio/log_frequency.py
@@ -12,7 +12,7 @@ from dataclasses import dataclass
 import numpy as np
 
 
-def _midi_to_freq(midi: int) -> float:
+def _midi_to_freq(midi: float) -> float:
     return 440.0 * 2.0 ** ((midi - 69) / 12.0)
